requirements.txt
torch
torchvision
pandas
numpy
scikit-learn
matplotlib
pillow
tqdm
kagglehub
grad-cam